+++ scripts/tags.js
@@ -1,0 +1,7 @@
+const emailInput = document.querySelector("#email");
+const pwdInput = document.querySelector("#password");
+const eyeIcons = document.querySelectorAll(".eye-icon");
+const errorMsgs = document.querySelectorAll(".error-message");
+const loginForm = document.querySelector("#login-form");
+
+export { emailInput, pwdInput, eyeIcons, errorMsgs, loginForm };
